+++ src/main/java/com/ibm/dto/PlanetUpdateDTO.java
@@ -15,5 +15,5 @@ public class PlanetUpdateDTO {
     private String name;
     private Integer planetType;
     private Double mass;
-    private Double circumference;
+    private Integer circumference;
 }
